feat: hide security and best practice score if the project is archived
